[python] extract converter_types.cpp from python_converter.cpp
Move 6 type-annotation methods of python_converter into
src/python-frontend/converter/converter_types.cpp: get_type_from_annotation,
extract_non_none_type, infer_tuple_struct_from_value, extract_type_from_boolean_op,
wrap_in_optional, unwrap_optional_if_needed.

Pure code movement — no logic change.
